+++ param_serv/worker.py
@@ -98,7 +98,7 @@ class ConnectorThread(threading.Thread):
             reception_json = receive_json(self.conn)
             data_size = struct.unpack("i", brecv(self.conn, struct.calcsize("i")))[0]
             with self.db[name] as inner:
-                self.db[name].inner = np.fromstring(brecv(self.conn, data_size))
+                self.db[name].inner = np.fromstring(brecv(self.conn, data_size), dtype=np.int64)
 
         except Exception, err:
             print(">>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>")
